// solar-system/types/auth.ts
export type AuthProps = {
  id: string;
  image_url: string;
  type: string;
  form_title_en: string;
  form_title_bg: string;
  form_description_en: string;
  form_description_bg: string;
  hero_title_en: string;
  hero_title_bg: string;
  hero_description_start_en: string;
  hero_description_start_bg: string;
  hero_description_end_en: string;
  hero_description_end_bg: string;
};